[llvm] Only init LLVM once.
