+++ compiler_gym/envs/llvm/service/RunService.cc
@@ -4,10 +4,60 @@
 // LICENSE file in the root directory of this source tree.
 #include "compiler_gym/envs/llvm/service/LlvmSession.h"
 #include "compiler_gym/service/runtime/Runtime.h"
+#include "llvm/InitializePasses.h"
+#include "llvm/Support/TargetSelect.h"
 
 const char* usage = R"(LLVM CompilerGym service)";
 
 using namespace compiler_gym::runtime;
 using namespace compiler_gym::llvm_service;
 
-int main(int argc, char** argv) { createAndRunCompilerGymService<LlvmSession>(argc, argv, usage); }
+namespace {
+
+void initLlvm() {
+  llvm::InitializeAllTargets();
+  llvm::InitializeAllTargetMCs();
+  llvm::InitializeAllAsmPrinters();
+  llvm::InitializeAllAsmParsers();
+
+  // Initialize passes.
+  llvm::PassRegistry& Registry = *llvm::PassRegistry::getPassRegistry();
+  llvm::initializeCore(Registry);
+  llvm::initializeCoroutines(Registry);
+  llvm::initializeScalarOpts(Registry);
+  llvm::initializeObjCARCOpts(Registry);
+  llvm::initializeVectorization(Registry);
+  llvm::initializeIPO(Registry);
+  llvm::initializeAnalysis(Registry);
+  llvm::initializeTransformUtils(Registry);
+  llvm::initializeInstCombine(Registry);
+  llvm::initializeAggressiveInstCombine(Registry);
+  llvm::initializeInstrumentation(Registry);
+  llvm::initializeTarget(Registry);
+  llvm::initializeExpandMemCmpPassPass(Registry);
+  llvm::initializeScalarizeMaskedMemIntrinPass(Registry);
+  llvm::initializeCodeGenPreparePass(Registry);
+  llvm::initializeAtomicExpandPass(Registry);
+  llvm::initializeRewriteSymbolsLegacyPassPass(Registry);
+  llvm::initializeWinEHPreparePass(Registry);
+  llvm::initializeDwarfEHPreparePass(Registry);
+  llvm::initializeSafeStackLegacyPassPass(Registry);
+  llvm::initializeSjLjEHPreparePass(Registry);
+  llvm::initializePreISelIntrinsicLoweringLegacyPassPass(Registry);
+  llvm::initializeGlobalMergePass(Registry);
+  llvm::initializeIndirectBrExpandPassPass(Registry);
+  llvm::initializeInterleavedAccessPass(Registry);
+  llvm::initializeEntryExitInstrumenterPass(Registry);
+  llvm::initializePostInlineEntryExitInstrumenterPass(Registry);
+  llvm::initializeUnreachableBlockElimLegacyPassPass(Registry);
+  llvm::initializeExpandReductionsPass(Registry);
+  llvm::initializeWasmEHPreparePass(Registry);
+  llvm::initializeWriteBitcodePassPass(Registry);
+}
+
+}  // anonymous namespace
+
+int main(int argc, char** argv) {
+  initLlvm();
+  createAndRunCompilerGymService<LlvmSession>(argc, argv, usage);
+}
